Rename internal Stack XRs to Backend
The internal XRs that install inference backends on remote clusters were
named KServeStack and DynamoStack. The public API calls these "backends"
(InferenceEnvironment.spec.backend: KServe | Dynamo), and the design doc
refers to them as backends throughout. "Stack" was a leftover from early
implementation when the concern was "install a stack of Helm charts."

This commit renames KServeStack to KServeBackend and DynamoStack to
DynamoBackend, aligning the internal XR names with the vocabulary used
everywhere else in the project. Directory names, composed resource keys,
Python method and variable names, and test fixtures are all updated to
match.

Fixes #17.
