+++ functions/compose-dynamo-backend/model
