+++ tests/test-kservebackend/main.py
@@ -1,5 +1,5 @@
 from .lib import resource as libresource
-from .model.ai.modelplane.infrastructure.dynamostack import v1alpha1 as dsv1alpha1
+from .model.ai.modelplane.infrastructure.kservebackend import v1alpha1 as kssv1alpha1
 from .model.io.crossplane.m.helm.providerconfig import v1beta1 as helmpcv1beta1
 from .model.io.crossplane.m.kubernetes.providerconfig import v1alpha1 as k8spcv1alpha1
 from .model.io.k8s.apimachinery.pkg.apis.meta import v1 as metav1
@@ -7,38 +7,45 @@ from .model.io.upbound.dev.meta.compositiontest import v1alpha1 as compositionte
 
 test = compositiontest.CompositionTest(
     metadata=metav1.ObjectMeta(
-        name="dynamostack-basic",
+        name="kservebackend-basic",
     ),
     spec=compositiontest.Spec(
-        compositionPath="apis/dynamostacks/composition.yaml",
-        xrPath="tests/test-dynamo-stack/xr.yaml",
-        xrdPath="apis/dynamostacks/definition.yaml",
+        compositionPath="apis/kservebackends/composition.yaml",
+        xrPath="tests/test-kservebackend/xr.yaml",
+        xrdPath="apis/kservebackends/definition.yaml",
         timeoutSeconds=120,
         validate=False,
         assertResources=[
             # Assert the XR spec is echoed back.
             libresource.model_to_dict(
-                dsv1alpha1.DynamoStack(
+                kssv1alpha1.KServeBackend(
                     metadata=metav1.ObjectMeta(
-                        name="dynamo-us-central-dynamo",
-                        namespace="modelplane-system",
+                        name="gpu-us-central1-kserve",
+                        namespace="gpu-us-central1",
                     ),
-                    spec=dsv1alpha1.Spec(
+                    spec=kssv1alpha1.Spec(
                         secrets=[
-                            dsv1alpha1.Secret(
+                            kssv1alpha1.Secret(
                                 type="Kubeconfig",
-                                name="dynamo-cluster-kubeconfig",
+                                name="gpu-us-central1-kubeconfig",
                                 key="kubeconfig",
+                            ),
+                            kssv1alpha1.Secret(
+                                type="GCPServiceAccountKey",
+                                name="gpu-us-central1-sa-key",
+                                key="private_key",
                             ),
                         ],
                     ),
                 )
             ),
             # Assert ProviderConfigs are composed on the first pass.
+            # Helm releases are gated on ProviderConfigs being observed,
+            # so they don't appear until the second reconcile.
             libresource.model_to_dict(
                 k8spcv1alpha1.ProviderConfig(
                     metadata=metav1.ObjectMeta(
-                        name="dynamo-us-central-dynamo-cluster",
+                        name="gpu-us-central1-kserve-cluster",
                         annotations={
                             "crossplane.io/composition-resource-name": "provider-config-kubernetes",
                         },
@@ -51,7 +58,7 @@ test = compositiontest.CompositionTest(
             libresource.model_to_dict(
                 helmpcv1beta1.ProviderConfig(
                     metadata=metav1.ObjectMeta(
-                        name="dynamo-us-central-dynamo-cluster",
+                        name="gpu-us-central1-kserve-cluster",
                         annotations={
                             "crossplane.io/composition-resource-name": "provider-config-helm",
                         },
